requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
chromadb>=1.4.0
ollama
pypdf
python-multipart
httpx
python-jose[cryptography]
passlib[bcrypt]
email-validator
